siges_backend_django/requirements.txt
Django~=5.2.1
gunicorn
djangorestframework
djangorestframework-simplejwt
